sanity: mount via sudo
"go test" is better called as normal user because then the build
environment and resulting files are for that user instead of root. But
usually mounting volumes will then fail because it is a privileged
operation and/or the necessary tools aren't in the default search path
of a normal user.

This can be avoided by wrapping the tools that might be needed (for
example, by the Kubernetes mount utility code) such that they get
invoked via sudo. If the CSI driver under test does not need them,
nothing changes, but if it does, the test now might succeed.

Fixes: kubernetes-csi/csi-test#73