sanity: Add starting_token check in ListVolumes
Adds test for expecting `ABORTED` gRPC code when an invalid
starting_token is passed to `Controller.ListVolumes()` and when the
starting_token value is greater than the actual number of volumes.

Also adds `Controller.ListVolumes()` test to check for the change
in volume list when a volume is added and removed, and test to pass
`MaxEntries` and receive a `NextToken` and ensure paginated list
volumes works.