Use custom gomock.Matcher for protobuf messages
The default gomock.Eq matcher uses reflect.DeepEqual to test for
equality, but as of protobuf 1.1.0, this will not work. Instead,
proto.Equal() needs to be used. Thus, implement a custom
protobuf matcher to test for message equality.

Add additional test for ControllerPublishVolume, which was the RPC seen
to trigger the protobuf comparison failure.