pkg/sanity: use different volume names
Within the test suite, each name is now unique. That simplifies
debugging and when one test fails to remove its volume, the other
tests are not affected. It also allows running tests in parallel.

In addition, a random suffix gets added to also prevent collisions
between different test runs. The random suffix gets reused for tests
running inside the same process, but will be different for a single
test run when running with "gingko -p".

Related-to: kubernetes-csi/csi-test#89.