pkg/sanity: reorganize suite
Instead of testing individual functions of each service, the
reorganized suite tests three services and for each service the
functions that it provides. This allows removing code duplication in
the common BeforeEach functions and enables further
enhancements (common cleanup code, not duplicating the same test
multiple times).

While at it, the term "services" is now used consistently instead of
"server", because that is what the CSI spec describes.