Add driver roles to sanity tests
This patch adds the ability to specify driver roles (Identity,
Controller, Node) for a given test run. This allows for a convenient,
code-based mechanism to parse which sanity tests will be run for a given
driver. This then allows for simpler "go test" CLI calls that helps
facilitate development and automation complexity.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>