Add target path checks and rm paths at teardown
This adds target path exists checks in the mock driver in
NodeStageVolume and NodePublishVolume to ensure that the paths actually
exist.

Also adds path cleanup in the teardown. This ensures that every test
creates and deletes their target directories. Without this, the test
setup function for Controller only tests created the target directory
and the same was being used by subsequent tests.