e2e.sh: simplify cleanup
All cleanup actions are now done via EXIT traps in subshells. That
removes the need to capture return codes.

The "Killed" message for the mock driver is removed, it looked odd to
run the script and have it end with that message.

Previously, "make test" when invoked from another script hang with
some mock driver processes still running. These changes fix that.