grpc: detect dead TCP connections
When testing against a driver that is installed elsewhere, for example
inside a virtual machine, we have to use TCP. When that virtual
machine dies without closing the TCP connection, for example because
of a simulated power loss, then gRPC by default only notices the dead
connection when the OS times out the TCP connection, which takes over
10 minutes (on Linux).

The gRPC documentation makes it sound like the keepalive pings are
only needed when there is no active method call, but in practice (with
gRPC v1.19.1 on Linux) such calls hung for over 10 minutes. Activating
the keepalive ping fixed that.