sanity: export several more items
It would be nice to build upon the test set and connection handling
that is already done by csi-sanity when including csi-sanity in a
custom test suite. The only reason why that wasn't possible is that
some of the functionality unnecessarily wasn't marked for exporting.

For an example that depends on these changes, see
https://github.com/intel/pmem-CSI/pull/237.