mock: support ephemeral inline volumes in Kubernetes 1.16
The upcoming Kubernetes 1.16 will have a special mode ("ephemeral
inline volumes") where it calls only NodePublishVolume for a
volume. The driver can detect that it is used in that mode by looking
at the csi.storage.k8s.io/ephemeral key in the volume context.

The csi-mock driver supports that mode and then relaxes some checks.