mock: support TCP
Sometimes it may be useful to connect to a CSI driver via TCP, for
example when controller and node service aren't running on the same
host. The mock driver can be used to test this mode of operation
now. It accepts the same "tcp://<host>:<port" and "unix://<path>"
syntax as other CSI drivers while still supporting just plain "<path>"
for backwards compatibility.