mock: avoid "not found" error in ControllerUnpublishVolume
The behavior of the external-attacher recently changed
(https://github.com/kubernetes-csi/external-attacher/pull/165) such
that it now treats "not found" as real error.

The effect was that some Kubernetes E2E tests (like "CSI mock volume
CSI workload information using mock driver should not be passed when
podInfoOnMount=false") sometimes ran for over 2 minutes, just waiting
for detatch. That the test then proceeds without marking the test as
failed is a bug in the test cleanup code which will be
fixed.

This slowdown is not deterministic: sometimes the detach is done early
enough while the volume still exists.

With this change, the same test completes in under 30 seconds.