Improve ListSnapshot filtering tests
We provide tests on ListSnapshots for filtering by snapshot IDs and
source volume IDs. However, the tests do not cover the case were
snapshot listing functionality exists in general but is missing
filtering capabilities: for this scenario the tests yield false
positives because the fixtures set up a single snapshot only.

We improve the tests by creating a number of snapshots of which a subset
is expected to be found.