sanity: test `NOT_FOUND` on `NodePublishVolume`
According to the CSI spec, a plugin must return a `NOT_FOUND` error when
`NodePublishVolume` is called with a `volume_id` of a volume that does
not (or no longer?) exist.

The `sanity` package has a test asserting this `NOT_FOUND` is returned
on `NodeUnpublishVolume`, but not for `NodePublishVolume`. This patch
adds this for completeness.

See: https://github.com/container-storage-interface/spec/blob/87a27a2496565857ed564284eaa762587feb9453/spec.md#nodepublishvolume-errors