Actually test CreateVolume with no capacity
The TODO comment is valid: the CSI spec says capacity is optional, but
that behavior when omitting it is up to the implementation.  However,
the code for "NoCapacity" was testing the same thing as "WithCapacity
1Gi" (the default csi-sanity size if the user didn't specify
differently), rather than matching the stated test of no capacity.
Allow for failures that depend on the implementation (we may need to
widen the list of allowable failures in the future).

Signed-off-by: Eric Blake <eblake@redhat.com>